JustGainsAPI.Tests/UsersSocialMediaAccountsControllerTest.cs
// <copyright file="UsersSocialMediaAccountsControllerTest.cs" company="APIMatic">
// Copyright (c) APIMatic. All rights reserved.
// </copyright>
using System;
using System.Collections.Generic;
using System.Globalization;
using System.IO;
using System.Threading.Tasks;
using APIMatic.Core.Utilities;
using JustGainsAPI.Standard;
using JustGainsAPI.Standard.Controllers;
using JustGainsAPI.Standard.Exceptions;
using JustGainsAPI.Standard.Http.Client;
using JustGainsAPI.Standard.Http.Response;
using JustGainsAPI.Standard.Utilities;
using Newtonsoft.Json.Converters;
using NUnit.Framework;

namespace JustGainsAPI.Tests
{
    /// <summary>
    /// UsersSocialMediaAccountsControllerTest.
    /// </summary>
    [TestFixture]
    public class UsersSocialMediaAccountsControllerTest : ControllerTestBase
    {
        /// <summary>
        /// Controller instance (for all tests).
        /// </summary>
        private UsersSocialMediaAccountsController controller;

        /// <summary>
        /// Setup test class.
        /// </summary>
        [OneTimeSetUp]
        public void SetUpDerived()
        {
            this.controller = this.Client.UsersSocialMediaAccountsController;
        }

        /// <summary>
        /// Retrieve a list of Users|Social Media Accounts with optional name filter.
        /// </summary>
        /// <returns>A <see cref="Task"/> representing the asynchronous unit test.</returns>
        [Test]
        public async Task TestTestListUsersSocialMediaAccounts()
        {
            // Parameters for the API call
            string name = null;

            // Perform API call
            Standard.Models.SocialMediaPlatformListResponse result = null;
            try
            {
                result = await this.controller.ListUsersSocialMediaAccountsAsync(name);
            }
            catch (ApiException)
            {
            }

            // Test response code
            Assert.AreEqual(200, HttpCallBack.Response.StatusCode, "Status should be 200");

            // Test headers
            Dictionary<string, string> headers = new Dictionary<string, string>();
            headers.Add("Content-Type", "application/json");

            Assert.IsTrue(
                    TestHelper.AreHeadersProperSubsetOf (
                    headers,
                    HttpCallBack.Response.Headers),
                    "Headers should match");
        }

        /// <summary>
        /// Check if a social media username is already taken.
        /// </summary>
        /// <returns>A <see cref="Task"/> representing the asynchronous unit test.</returns>
        [Test]
        public async Task TestTestValidateSocialMediaUsername()
        {
            // Parameters for the API call
            string socialMediaPlatformCode = "INSTAGRAM";
            string username = "fitness_guru";

            // Perform API call
            Standard.Models.SocialMediaPlatformsValidateUsernameResponse result = null;
            try
            {
                result = await this.controller.ValidateSocialMediaUsernameAsync(socialMediaPlatformCode, username);
            }
            catch (ApiException)
            {
            }

            // Test response code
            Assert.AreEqual(200, HttpCallBack.Response.StatusCode, "Status should be 200");

            // Test headers
            Dictionary<string, string> headers = new Dictionary<string, string>();
            headers.Add("Content-Type", "application/json");

            Assert.IsTrue(
                    TestHelper.AreHeadersProperSubsetOf (
                    headers,
                    HttpCallBack.Response.Headers),
                    "Headers should match");
        }
    }
}